레시피 수정 기능 개발중

+++ src/pages/RecipeDetailViewPage.js
@@ -30,7 +30,7 @@ function RecipeDetailViewPage() {
         const likeStatusResponse = await axios.get(
           `http://localhost:8080/recipes/${recipeId}/likes/status`,
           {
-            params: userId, // 현재 로그인한 사용자 ID
+            params: { userId }, // 현재 로그인한 사용자 ID
           }
         );
         setIsLiked(likeStatusResponse.data);
@@ -40,15 +40,28 @@ function RecipeDetailViewPage() {
     };
 
     fetchRecipeDetails();
-  }, [recipeId]);
+  }, [recipeId, userId]);
 
   const handleDelete = async () => {
-    try {
-      await axios.delete(`http://localhost:8080/recipes/${recipeId}`);
-      navigate("/recipes");
-    } catch (error) {
-      console.error("레시피 삭제 실패:", error);
+    // 로그인된 사용자의 ID와 레시피 작성자의 ID가 일치하는지 확인
+    if (userId && recipe.authorId === userId) {
+      const confirmDelete = window.confirm("정말로 레시피를 삭제하시겠습니까?");
+      if (confirmDelete) {
+        // 사용자가 '확인'을 클릭한 경우 삭제 로직 실행
+        try {
+          await axios.delete(`http://localhost:8080/recipes/${recipeId}`);
+          navigate("/recipes");
+        } catch (error) {
+          console.error("레시피 삭제 실패:", error);
+        }
+      }
+    } else {
+      alert("본인의 레시피만 삭제할 수 있습니다.");
     }
+  };
+
+  const navigateToEdit = () => {
+    navigate(`/edit-recipe/${recipeId}`);
   };
 
   const toggleLike = async () => {
@@ -83,7 +96,12 @@ function RecipeDetailViewPage() {
       <p>{recipe.userNickname}</p>
       <p>{formatDate(recipe.registerTime)}</p>
       <h2>{recipe.recipeName}</h2>
-      <button onClick={handleDelete}>삭제</button>
+      {userId && recipe.authorId === userId && (
+        <>
+          <button onClick={handleDelete}>삭제</button>
+          <button onClick={navigateToEdit}>수정</button>
+        </>
+      )}
       <img
         src={`http://localhost:8080/recipeimgs/${recipe.photoUrl}`}
         alt={recipe.recipeName}
